// frontend/Open2Work/src/interfaces/loginInterface.ts
// Generated by https://quicktype.io

export interface UserLogued {
    message: string;
    token:   string;
    user:    User;
    client?: User;

}

export interface User {
    isDev:       boolean;
    oldTeams:    any[];
    _id:         string;
    name:        string;
    surname:     string;
    email:       string;
    password:    string;
    role:        string;
    avatar:      string;
    social:      Social;
    info:        Info;
    __v:         number;
    currentTeam: string;
    orders: OrderElement[];
}

export interface Social {
    linkedin:  string;
    portfolio: string;
    github:    string;
    _id:       string;
}

export interface Info {
    organization: string;
    time_availability: string;
    time_zone:         string;
    experience:        number;
    language:          string[];
    _id:               string;
}

export interface OrderElement {
    _id:         string;
    client:      string;
    team:        string;
    devs_ok:     any[];
    devs_not:    any[];
    description: string;
    createdAt:   string;
    __v:         number;
}